requirements.txt
aws-cdk-lib==2.46.0
constructs>=10.0.0,<11.0.0
boto3>=1.24
black>=22.10
isort>=5.10
